--- .storybook/preview.ts
@@ -1,16 +1,0 @@
-import "@/mds/style.linaria.global";
-import { Preview } from "@storybook/react";
-
-const preview: Preview = {
-  parameters: {
-    actions: { argTypesRegex: "^on[A-Z].*" },
-    controls: {
-      matchers: {
-        color: /(background|color)$/i,
-        date: /Date$/
-      }
-    }
-  }
-};
-
-export default preview;
